update jpa view manager

+++ src/main/java/com/quynhtd/source_code_final/service/OrderService.java
@@ -1,7 +1,7 @@
 package com.quynhtd.source_code_final.service;
 
-
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,22 +11,34 @@ import com.quynhtd.source_code_final.repository.OrderRepository;
 
 @Service
 public class OrderService {
-    @Autowired
-    private OrderRepository orderRepository;
+	  @Autowired
+	    private OrderRepository orderRepository;
 
-    public List<Order> getAllOrders() {
-        return orderRepository.findAll();
-    }
+	    public List<Order> getAllOrders() {
+	        return orderRepository.findAll();
+	    }
 
-    public Order getOrderById(Long id) {
-        return orderRepository.findById(id).orElse(null);
-    }
+	    public Optional<Order> getOrderById(Long id) {
+	        return orderRepository.findById(id);
+	    }
 
-    public Order saveOrder(Order order) {
-        return orderRepository.save(order);
-    }
+	    public Order createOrder(Order Order) {
+	        return orderRepository.save(Order);
+	    }
 
-    public void deleteOrder(Long id) {
-        orderRepository.deleteById(id);
-    }
-}
+	    public Order updateOrder(Long id, Order order) {
+	        if (orderRepository.existsById(id)) {
+	            order.setId(id);
+	            return orderRepository.save(order);
+	        }
+	        return null;
+	    }
+
+	    public boolean deleteOrder(Long id) {
+	        if (orderRepository.existsById(id)) {
+	            orderRepository.deleteById(id);
+	            return true;
+	        }
+	        return false;
+	    }
+	}
